+++ src/hooks/useSearch.js
@@ -1,0 +1,6 @@
+import { useContext } from 'react'
+import { SearchContext } from '../features/search/context/SearchContext'
+
+export default function useSearch() {
+    return useContext(SearchContext)
+}
