+++ src/layouts/Container.jsx
@@ -4,16 +4,19 @@ import { Outlet } from "react-router-dom";
 import Footer from "./Footer";
 import LiveChat from "../pages/LiveChat";
 import ChatContextProvider from "../features/livechat/contexts/ChatContext";
+import SearchContextProvider from "../features/search/context/SearchContext";
 
 export default function Wrapper() {
   return (
     <>
-      <Navbar />
-      <ChatContextProvider>
-        <LiveChat />
-      </ChatContextProvider>
-      <Outlet />
-      <Footer />
+      <SearchContextProvider>
+        <Navbar />
+        <ChatContextProvider>
+          <LiveChat />
+        </ChatContextProvider>
+        <Outlet />
+        <Footer />
+      </SearchContextProvider>
     </>
   );
 }
